Accelerate fetch of .svg with dek_fetch.py (Linux)
The file of addresses to consult now may contain comment lines
after a # sign.  The harvest of .svg in Linux (detected by the
script) attempts to benefit from the more powerful wget2, while
retaing wget as fallback.  With xargs, the collection of .svg
files is parallelized to use (n - 1) accessible processor cores.

The collection of .svg remains strictly linear and limted to wget
when using Windows, or MacOS (not tested).