split svg optimizing moderator into two, concept
Seeking a strategy to optimize the simplification of the .svg
for a smaller file size of individual .svg, as well for the
data set as such (about 40k individual files), there should
be a better way doing this.  Departing from dek_optimize.sh,
two simpler scripts for the bash shell were written which are
to be tested for processing the .svg both reliably, as well as
hopefully higher rate than their ancestor in common.