simplify tag assignment, white list and dimensions test
The organization of the symbolizations on Wikipedia / wikimedia
changed, some of them are categorized which can be used to
assign tags about dedicated (topical) sub sets.  This is why
the hypenator-based approach (for now) is not used, and testing
criteria now focus on un-categorized DEK plates which equally
have to be of a uniform dimension.

If this working well enough, extension to the symbolization with
multiple words, about geography, etc may be incrementally added.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>