adjust scripts to local needs
There no longer is a clear track in sequence in changes leading
to the present state, a result of multiple tests.  A few of the
symbolizations will be dropped, either for dimension, or (maybe)
because of some odds in the definition of the .svg which seem to
represent an obstacle for the svgcleaner moderated by optimizer5b.
However, the rate of drop-outs, and overall processing rate is
acceptable.